refactored services to have separte interface and implementation
